refactor(discord): extract watcher streaming status tick
Move the throttled streaming status tick anchored at last_status_update.elapsed() through the streaming placeholder edit/persist block into tmux_watcher/streaming_status_tick.rs. The moved slice covers orphan reclaim, streaming suppression, status-panel create/bind, rollover, re-anchor, and placeholder edit.

Loop-control mapping: 5 direct continue branches map to StreamingStatusTickOutcome::ContinueStreamingLoop; 0 enclosing-loop break branches map out. The 6 rollover-local break branches remain internal to the moved loop.

Threaded locals: 30 total via context/state structs (15 read-only inputs, 15 mutable state locals). Mutable state is committed back through the tick-state macro before every mapped continue and at fallthrough.

Maintenance gates: python3 scripts/generate_inventory_docs.py; python3 scripts/check_agent_maintenance_docs.py; bash scripts/ci-script-checks.sh; cargo fmt. tmux_watcher.rs ratcheted to 5975 raw/prod lines; streaming_status_tick.rs is 963 lines with a file-specific namespace-cap admission.
